\name{SplicingTypesAnno-package}
\Rdversion{0.99}
\alias{SplicingTypesAnno-package}
\alias{SplicingTypesAnno}


\docType{package}
\title{
Annotation tool for alternative splicing types by RNA-Seq data
}
\description{
This package annotates four major splicing types by RNA-Seq data, 
including intron retention, exon skipping, alternative donor site, 
and alternative acceptor site. 
As a post-processing tool for alignment files, it annotates the alternative splicing types with 
details at the intron or exon level.
}
\details{

  See \code{packageDescription('SplicingTypesAnno')} for package details. 

}
\author{
Xiaoyong Sun
Maintainer: johnsunx1<johnsunx1@gmail.com>
}


\keyword{ package }
\examples{
packageDescription('SplicingTypesAnno')
}
