\name{updateTRUEIntron}
\Rdversion{0.99}

\alias{updateTRUEIntron}
\alias{updateTRUEIntron,character-method}

\title{Splicing type function}

\description{
Splicing type function.
}

\usage{

}

\arguments{

}


\value{

}


\author{X. Sun}

\seealso{

}

\examples{

}

\keyword{methods}
\keyword{classes}
